Add anagrams example/benchmark
The anagrams-hashmap program is an implementation of "Letterpress
Cheating", a program to find all the words that can be built from a bag
of letters. This implementation looks for the 7000-odd results from a
specific string, using different hashers and reporting the time.

Lots of type-hackery to parameterize the functions by Hashers.